requirements.txt
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
docxtpl>=0.17.0
jinja2>=3.1.0
pyyaml>=6.0
mistune>=3.0.0
